幸存者模式 新增障碍物阻挡功能，新增雷击球

+++ Assets/Scripts/HotFix/_Gameplay/AWaves/WaveRewardManager.cs
@@ -314,14 +314,14 @@ namespace MoreMountains
             if (player == null || player.Health == null) 
                 return;
 
-            float healAmount;
+            int healAmount;
             if (reward.value >= 100)
             {
                 healAmount = player.maxHealth;
             }
             else
             {
-                healAmount = player.maxHealth * (reward.value / 100f);
+                healAmount = (int)(player.maxHealth * (reward.value / 100f));
             }
 
             player.Health.ReceiveHealth(new Heal { Value = healAmount }, null, player);
