+++ Assets/Scripts/HotFix/_TopDownEngine/Characters/Buff/Buffs/Buff_Rune_Conqueror.cs
@@ -22,7 +22,7 @@ namespace MoreMountains
             var damage = e.Dmg.DamageDealt;
             var coeff = HealCoeff.Value(this);
             var heal = damage * coeff;
-            Target.Health.ReceiveHealth(Heal.Fixed(heal));
+            Target.Health.ReceiveHealth(Heal.Fixed((int)heal));
         }
     }
 }
