+++ Assets/Scripts/HotFix/_TopDownEngine/Characters/Buff/UniGAS/GameplayEffect/Buff.cs
@@ -145,7 +145,7 @@ namespace MoreMountains
 
                 if (value > 0)
                 {
-                    var dmg = new Dmg(value, mag.DmgType, mag.DmgAlgo);
+                    var dmg = new Dmg((int)value, mag.DmgType, mag.DmgAlgo);
                     Target.Health.Damage(ref dmg, gameObject, source: Source.Character);
                 }
             }
@@ -156,7 +156,7 @@ namespace MoreMountains
                 var value = mag.Value(this);
                 if (value > 0)
                 {
-                    var heal = new Heal(value, mag.HealAlgo);
+                    var heal = new Heal((int)value, mag.HealAlgo);
                     Target.Health.ReceiveHealth(heal, source: Target.Character);
                 }
             }
